Add fn Editor::handle_event
