+++ crates/kas-widgets/src/edit/edit_field.rs
@@ -6,12 +6,10 @@
 //! The [`EditField`] and [`EditBox`] widgets, plus supporting items
 
 use super::*;
-use kas::event::components::TextInputAction;
-use kas::event::{CursorIcon, ElementState, FocusSource};
-use kas::event::{Ime, ImePurpose};
+use kas::event::CursorIcon;
 use kas::messages::{ReplaceSelectedText, SetValueText};
 use kas::prelude::*;
-use kas::text::{Effect, EffectFlags, NotReady};
+use kas::text::{Effect, EffectFlags};
 use kas::theme::TextClass;
 use std::fmt::{Debug, Display};
 use std::str::FromStr;
@@ -216,283 +214,7 @@ mod EditField {
         }
 
         fn handle_event(&mut self, cx: &mut EventCx, data: &G::Data, event: Event) -> IsUsed {
-            let action = match event {
-                Event::NavFocus(source) if source == FocusSource::Key => {
-                    if !self.input_handler.is_selecting() {
-                        self.request_key_focus(cx, source);
-                    }
-                    CmdAction::Used
-                }
-                Event::NavFocus(_) => CmdAction::Used,
-                Event::LostNavFocus => CmdAction::Used,
-                Event::SelFocus(source) => {
-                    // NOTE: sel focus implies key focus since we only request
-                    // the latter. We must set before calling self.set_primary.
-                    self.has_key_focus = true;
-                    if source == FocusSource::Pointer {
-                        self.set_primary(cx);
-                    }
-
-                    CmdAction::Used
-                }
-                Event::KeyFocus => {
-                    self.has_key_focus = true;
-                    self.set_view_offset_from_cursor(cx);
-
-                    if self.current.is_none() {
-                        let hint = Default::default();
-                        let purpose = ImePurpose::Normal;
-                        let surrounding_text = self.ime_surrounding_text();
-                        cx.replace_ime_focus(self.id.clone(), hint, purpose, surrounding_text);
-                        CmdAction::FocusGained
-                    } else {
-                        CmdAction::Used
-                    }
-                }
-                Event::LostKeyFocus => {
-                    self.has_key_focus = false;
-                    cx.redraw();
-                    if !self.current.is_ime_enabled() {
-                        CmdAction::FocusLost
-                    } else {
-                        CmdAction::Used
-                    }
-                }
-                Event::LostSelFocus => {
-                    // NOTE: we can assume that we will receive Ime::Disabled if IME is active
-                    if !self.selection.is_empty() {
-                        self.save_undo_state(None);
-                        self.selection.set_empty();
-                    }
-                    self.input_handler.stop_selecting();
-                    cx.redraw();
-                    CmdAction::Used
-                }
-                Event::Command(cmd, code) => match self.editor.cmd_action(cx, cmd, code) {
-                    Ok(action) => action,
-                    Err(NotReady) => CmdAction::Used,
-                },
-                Event::Key(event, false) if event.state == ElementState::Pressed => {
-                    if let Some(text) = &event.text {
-                        self.save_undo_state(Some(EditOp::KeyInput));
-                        if self.received_text(cx, text) == Used {
-                            CmdAction::Edit
-                        } else {
-                            CmdAction::Unused
-                        }
-                    } else {
-                        let opt_cmd = cx
-                            .config()
-                            .shortcuts()
-                            .try_match_event(cx.modifiers(), event);
-                        if let Some(cmd) = opt_cmd {
-                            match self.editor.cmd_action(cx, cmd, Some(event.physical_key)) {
-                                Ok(action) => action,
-                                Err(NotReady) => CmdAction::Used,
-                            }
-                        } else {
-                            CmdAction::Unused
-                        }
-                    }
-                }
-                Event::Ime(ime) => match ime {
-                    Ime::Enabled => {
-                        match self.current {
-                            CurrentAction::None => {
-                                self.current = CurrentAction::ImeStart;
-                                self.set_ime_cursor_area(cx);
-                            }
-                            CurrentAction::ImeStart | CurrentAction::ImePreedit { .. } => {
-                                // already enabled
-                            }
-                            CurrentAction::Selection => {
-                                // Do not interrupt selection
-                                cx.cancel_ime_focus(self.id_ref());
-                            }
-                        }
-                        if !self.has_key_focus {
-                            CmdAction::FocusGained
-                        } else {
-                            CmdAction::Used
-                        }
-                    }
-                    Ime::Disabled => {
-                        self.clear_ime();
-                        if !self.has_key_focus {
-                            CmdAction::FocusLost
-                        } else {
-                            CmdAction::Used
-                        }
-                    }
-                    Ime::Preedit { text, cursor } => {
-                        self.save_undo_state(None);
-                        let mut edit_range = match self.current.clone() {
-                            CurrentAction::ImeStart if cursor.is_some() => self.selection.range(),
-                            CurrentAction::ImeStart => return Used,
-                            CurrentAction::ImePreedit { edit_range } => edit_range.cast(),
-                            _ => return Used,
-                        };
-
-                        self.text.replace_range(edit_range.clone(), text);
-                        edit_range.end = edit_range.start + text.len();
-                        if let Some((start, end)) = cursor {
-                            self.selection.set_sel_index_only(edit_range.start + start);
-                            self.selection.set_edit_index(edit_range.start + end);
-                        } else {
-                            self.selection.set_cursor(edit_range.start + text.len());
-                        }
-
-                        self.current = CurrentAction::ImePreedit {
-                            edit_range: edit_range.cast(),
-                        };
-                        self.edit_x_coord = None;
-                        self.prepare_and_scroll(cx, false);
-                        CmdAction::Used
-                    }
-                    Ime::Commit { text } => {
-                        self.save_undo_state(Some(EditOp::Ime));
-                        let edit_range = match self.current.clone() {
-                            CurrentAction::ImeStart => self.selection.range(),
-                            CurrentAction::ImePreedit { edit_range } => edit_range.cast(),
-                            _ => return Used,
-                        };
-
-                        self.text.replace_range(edit_range.clone(), text);
-                        self.selection.set_cursor(edit_range.start + text.len());
-
-                        self.current = CurrentAction::ImePreedit {
-                            edit_range: self.selection.range().cast(),
-                        };
-                        self.edit_x_coord = None;
-                        self.prepare_and_scroll(cx, false);
-                        CmdAction::Edit
-                    }
-                    Ime::DeleteSurrounding {
-                        before_bytes,
-                        after_bytes,
-                    } => {
-                        self.save_undo_state(None);
-                        let edit_range = match self.current.clone() {
-                            CurrentAction::ImeStart => self.selection.range(),
-                            CurrentAction::ImePreedit { edit_range } => edit_range.cast(),
-                            _ => return Used,
-                        };
-
-                        if before_bytes > 0 {
-                            let end = edit_range.start;
-                            let start = end - before_bytes;
-                            if self.as_str().is_char_boundary(start) {
-                                self.text.replace_range(start..end, "");
-                                self.selection.delete_range(start..end);
-                            } else {
-                                log::warn!("buggy IME tried to delete range not at char boundary");
-                            }
-                        }
-
-                        if after_bytes > 0 {
-                            let start = edit_range.end;
-                            let end = start + after_bytes;
-                            if self.as_str().is_char_boundary(end) {
-                                self.text.replace_range(start..end, "");
-                            } else {
-                                log::warn!("buggy IME tried to delete range not at char boundary");
-                            }
-                        }
-
-                        if let Some(text) = self.ime_surrounding_text() {
-                            cx.update_ime_surrounding_text(self.id_ref(), text);
-                        }
-
-                        CmdAction::Used
-                    }
-                },
-                Event::PressStart(press) if press.is_tertiary() => {
-                    match press.grab_click(self.id()).complete(cx) {
-                        Unused => CmdAction::Unused,
-                        Used => CmdAction::Used,
-                    }
-                }
-                Event::PressEnd { press, .. } if press.is_tertiary() => {
-                    self.set_cursor_from_coord(cx, press.coord);
-                    self.cancel_selection_and_ime(cx);
-                    self.request_key_focus(cx, FocusSource::Pointer);
-
-                    if let Some(content) = cx.get_primary() {
-                        self.save_undo_state(Some(EditOp::Clipboard));
-
-                        let index = self.selection.edit_index();
-                        let range = self.trim_paste(&content);
-
-                        self.text
-                            .replace_range(index..index, &content[range.clone()]);
-                        self.selection.set_cursor(index + range.len());
-                        self.edit_x_coord = None;
-                        self.prepare_and_scroll(cx, false);
-
-                        CmdAction::Edit
-                    } else {
-                        CmdAction::Used
-                    }
-                }
-                event => match self.editor.input_handler.handle(cx, self.core.id(), event) {
-                    TextInputAction::Used => CmdAction::Used,
-                    TextInputAction::Unused => CmdAction::Unused,
-                    TextInputAction::PressStart {
-                        coord,
-                        clear,
-                        repeats,
-                    } => {
-                        if self.current.is_ime_enabled() {
-                            self.clear_ime();
-                            cx.cancel_ime_focus(self.id_ref());
-                        }
-                        self.save_undo_state(Some(EditOp::Cursor));
-                        self.current = CurrentAction::Selection;
-
-                        self.set_cursor_from_coord(cx, coord);
-                        self.selection.set_anchor(clear);
-                        if repeats > 1 {
-                            self.editor
-                                .selection
-                                .expand(&self.editor.text, repeats >= 3);
-                        }
-
-                        self.request_key_focus(cx, FocusSource::Pointer);
-                        CmdAction::Used
-                    }
-                    TextInputAction::PressMove { coord, repeats } => {
-                        if self.current == CurrentAction::Selection {
-                            self.set_cursor_from_coord(cx, coord);
-                            if repeats > 1 {
-                                self.editor
-                                    .selection
-                                    .expand(&self.editor.text, repeats >= 3);
-                            }
-                        }
-
-                        CmdAction::Used
-                    }
-                    TextInputAction::PressEnd { coord } => {
-                        if self.current.is_ime_enabled() {
-                            self.clear_ime();
-                            cx.cancel_ime_focus(self.id_ref());
-                        }
-                        self.save_undo_state(Some(EditOp::Cursor));
-                        if self.current == CurrentAction::Selection {
-                            self.set_primary(cx);
-                        } else {
-                            self.set_cursor_from_coord(cx, coord);
-                            self.selection.set_empty();
-                        }
-                        self.current = CurrentAction::None;
-
-                        self.request_key_focus(cx, FocusSource::Pointer);
-                        CmdAction::Used
-                    }
-                },
-            };
-
-            match action {
+            match self.editor.handle_event(cx, event) {
                 CmdAction::Unused => Unused,
                 CmdAction::Used | CmdAction::Cursor => Used,
                 CmdAction::FocusGained => {
